feat(plugin-gis-viewer): CB-664 remove extra GISPresentation loader component
