+++ webapp/packages/plugin-gis-viewer/src/GISViewer.tsx
@@ -7,13 +7,24 @@
  */
 
 import type { TabContainerPanelComponent } from '@cloudbeaver/core-blocks';
+import { ComplexLoader, Loader } from '@cloudbeaver/core-blocks';
 import type { IDataValuePanelProps, IDatabaseResultSet } from '@cloudbeaver/plugin-data-viewer';
 
-import { GISValuePresentationLoader } from './GISValuePresentationLoader';
+async function loader() {
+  const { GISValuePresentation } = await import('./GISValuePresentation');
+  return { GISValuePresentation };
+}
 
 export const GISViewer: TabContainerPanelComponent<IDataValuePanelProps<any, IDatabaseResultSet>> = function GISViewer({ model, resultIndex }) {
   return (
-    <GISValuePresentationLoader model={model} resultIndex={resultIndex} />
+    <ComplexLoader
+      loader={loader}
+      placeholder={<Loader />}
+    >
+      {({ GISValuePresentation }) => (
+        <GISValuePresentation model={model} resultIndex={resultIndex} />
+      )}
+    </ComplexLoader>
   );
 }
 ;
